pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "psd2svg"
version = "0.3.0"
description = "Convert PSD file to SVG file"
authors = [
    { name = "Kota Yamaguchi", email = "yamaguchi_kota@cyberagent.co.jp" },
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.10,<3.15"
keywords = ["photoshop", "psd", "svg"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Multimedia :: Graphics :: Viewers",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
]
dependencies = [
    "pillow",
    "svgwrite", 
    "numpy",
    "psd-tools>=1.10",
]

[dependency-groups]
dev = [
    "pytest",
    "imagehash",
    "mypy",
    "ruff",
    "ipykernel>=6.30.1",
]

[project.urls]
Homepage = "https://github.com/kyamagu/psd2svg"
Repository = "https://github.com/kyamagu/psd2svg"

[project.scripts]
psd2svg = "psd2svg.__main__:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
